evaluation/pyproject.toml
[project]
name = "evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "scipy>=1.16.0",
    "matplotlib>=3.7.0",
    "pyqt5>=5.15.11",
]
